Rename vectordb to vectordb_pinecone
