+++ collections/collections.go
@@ -5,7 +5,7 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/pzierahn/chatbot_services/auth"
 	pb "github.com/pzierahn/chatbot_services/proto"
-	"github.com/pzierahn/chatbot_services/vectordb"
+	"github.com/pzierahn/chatbot_services/vectordb_pinecone"
 )
 
 const (
@@ -17,10 +17,10 @@ type Service struct {
 	auth     auth.Service
 	db       *pgxpool.Pool
 	storage  *storage.BucketHandle
-	vectorDB *vectordb.DB
+	vectorDB *vectordb_pinecone.DB
 }
 
-func NewServer(auth auth.Service, db *pgxpool.Pool, storage *storage.BucketHandle, vectorDB *vectordb.DB) *Service {
+func NewServer(auth auth.Service, db *pgxpool.Pool, storage *storage.BucketHandle, vectorDB *vectordb_pinecone.DB) *Service {
 	return &Service{
 		db:       db,
 		storage:  storage,
